Refactored action files/folders and added actions for fetching tasks
