+++ client/src/actions/types.js
@@ -1,2 +1,7 @@
+// Action types for authentication
 export const AUTH_USER = "auth_user";
 export const AUTH_ERROR = "auth_error";
+
+// Action types for tasks
+export const FETCH_TASKS = "fetch_tasks";
+export const FETCH_FAILED = "fetch_failed";
