+++ client/src/reducers/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from "redux";
 import { reducer as formReducer } from "redux-form";
 import auth from "./auth";
+import task from "./task";
 
 export default combineReducers({
   auth,
+  task,
   form: formReducer
 });
